+++ framework/json-db/src/main/java/is/codion/framework/json/db/CountDeserializer.java
@@ -54,7 +54,8 @@ final class CountDeserializer extends StdDeserializer<Count> {
 		JsonNode havingNode = jsonNode.get("having");
 		Condition having = entityObjectMapper.deserializeCondition(definition, havingNode);
 
-		return Count.builder(where)
+		return Count.builder()
+						.where(where)
 						.having(having)
 						.build();
 	}