+++ framework/json-db/src/test/java/is/codion/framework/json/db/DatabaseObjectMapperTest.java
@@ -108,7 +108,8 @@ public final class DatabaseObjectMapperTest {
 
 	@Test
 	void count() throws JsonProcessingException {
-		Count count = Count.builder(Department.DEPTNO.between(1, 2))
+		Count count = Count.builder()
+						.where(Department.DEPTNO.between(1, 2))
 						.having(Department.NAME.equalTo("TEST"))
 						.build();