+++ changelog.md
@@ -28,6 +28,7 @@ Codion Change Log
 - Unit tests improved.
 ### is.codion.framework.db.core
 - Unit tests improved.
+- EntityConnection.Count.Builder.WhereStage added.
 ### is.codion.framework.db.http
 - HttpEntityConnection.DISCONNECT_ON_CLOSE configuration value added.
 ### is.codion.framework.db.local